+++ src/components/Builder/SearchInput.js
@@ -1,65 +1,41 @@
 import {
   useCallback,
-  useEffect,
   useRef,
   useState,
-  useMemo,
 } from 'react';
 import PropTypes from 'prop-types';
-import Fuse from 'fuse.js/dist/fuse.common';
 import SearchIcon from '../../assets/search.svg';
 import CrossIcon from '../../assets/svg/close.svg';
 import { useTranslatedTexts } from '../../utils/hooks';
 
 const SearchInput = ({
-  elements = [],
-  searchKeys = [],
-  setElements = () => {},
+  onSearch,
 }) => {
   const { SEARCH } = useTranslatedTexts();
   const searchInputRef = useRef(null);
   const [searchValue, setSearchValue] = useState('');
 
-  const searchEngine = useMemo(() => {
-    return new Fuse(elements, { keys: searchKeys, threshold: 0.4 });
-  }, [elements, searchKeys]);
+  const handleSearch = useCallback(e => {
+    const { value } = e?.target || {};
+    setSearchValue(value);
+    onSearch?.(value);
+  }, [onSearch]);
 
-  useEffect(() => {
-    setElements(elements);
+  const handleClearSearch = useCallback(() => {
     setSearchValue('');
-  }, [elements]);
-
-  const onSearchButtonClick = useCallback(() => {
-    if (searchValue) {
-      setSearchValue('');
-      setElements(elements);
-    }
+    onSearch?.('');
     if (searchInputRef.current) {
       searchInputRef.current.focus();
     }
-  }, [searchValue]);
-
-  const onSearchChange = useCallback(e => {
-    const _searchValue = e.target.value;
-    setSearchValue(_searchValue);
-
-    if (!_searchValue || !searchKeys) {
-      return setElements(elements);
-    }
-
-    const filteredElements = searchEngine.search(_searchValue).map(i => i.item);
-
-    setElements(filteredElements.length ? filteredElements : 'noResult');
-  }, [elements, searchKeys]);
+  }, [onSearch]);
 
   const handleKeyDown = useCallback(e => {
     e.stopPropagation();
     if (e.key === 'Escape' && searchInputRef.current) {
       setSearchValue('');
-      setElements(elements);
       searchInputRef.current.blur();
     }
-  }, [elements]);
+  }, []);
 
   return (
     <div className="jfReport-searchContainer">
@@ -67,23 +43,22 @@ const SearchInput = ({
         <input
           ref={searchInputRef}
           className="js-searchInput section-search-input"
-          onChange={onSearchChange}
+          onChange={handleSearch}
           onKeyDown={handleKeyDown}
           placeholder={SEARCH}
           type="text"
           value={searchValue}
         />
-        <button
-          className="section-search-button search-icon"
-          onClick={onSearchButtonClick}
-          type="button"
+        <span
+          className="search-icon"
+          style={{ left: '8px', position: 'absolute' }}
         >
           <SearchIcon className="section-search-icon" />
-        </button>
+        </span>
         {searchValue && (
           <button
             className="section-search-button search-delete"
-            onClick={onSearchButtonClick}
+            onClick={handleClearSearch}
             type="button"
           >
             <CrossIcon className="section-search-icon delete" />
@@ -95,12 +70,7 @@ const SearchInput = ({
 };
 
 SearchInput.propTypes = {
-  elements: PropTypes.arrayOf(PropTypes.shape({})),
-  searchKeys: PropTypes.arrayOf(PropTypes.shape({
-    name: PropTypes.string,
-    weight: PropTypes.number,
-  })),
-  setElements: PropTypes.func,
+  onSearch: PropTypes.func,
 };
 
 export default SearchInput;